feat: pre-render the landing page into index.html
The homepage shipped as an empty `#app` shell, so crawlers and agents that
do not run JavaScript saw 34 characters of text. Move the markup into a pure
string builder (`src/landing.ts`) and render it into `index.html` at build
time with a Vite `transformIndexHtml` plugin. `main.ts` now enhances the
pre-rendered markup and only renders when it is absent.

Raw homepage text goes from 34 to ~3,630 characters, with one `h1`,
sequential heading levels, and the existing `/docs` links crawlable.
